+++ public/controllers/agent/components/agents-welcome.js
