+++ public/controllers/management/components/management-welcome.js
