+++ public/controllers/overview/components/overview-welcome.js
