+++ public/controllers/overview/index.js
@@ -11,7 +11,7 @@
  */
 import { uiModules } from 'ui/modules';
 import { OverviewController } from './overview';
-import { WelcomeScreen } from './components/welcome';
+import { WelcomeScreen } from './components/overview-welcome';
 import { Stats } from './components/stats';
 import { AlertsStats } from './components/alerts-stats';
 import { RequirementCard } from './components/requirement-card';
