+++ public/templates/management/configuration/agent-configuration.pug
@@ -1,5 +1,5 @@
 include ./agent-configuration.head
-include ./welcome.html
+include ./configuration-welcome.html
 include ./global-configuration/global-configuration.pug
 include ./log-settings/log-settings.pug
 include ./client/client.html
