+++ public/templates/management/configuration/configuration.pug
@@ -1,5 +1,5 @@
 include ./configuration.head
-include ./welcome.html
+include ./configuration-welcome.html
 include ./global-configuration/global-configuration.pug
 include ./cluster/cluster.html
 include ./config-edition/config-edition.html
