+++ public/templates/management/management.pug
@@ -1,5 +1,5 @@
 include ./management.head
-include ./welcome.html
+include ./management-welcome.html
 //- include ./configuration/configuration.pug
 include ./monitoring/monitoring.pug
 include ./management.html
